Move shared code into common.py
